requirements.txt
chromadb
transformers<5
torch
torchvision
einops
numpy
Pillow
PyMuPDF
nltk
pyyaml
addict
safetensors
fastapi
uvicorn
python-multipart
img2pdf
tqdm
vllm

# Optional (recommended for Qwen3-VL image resize parity):
# qwen-vl-utils

# 可选：使用 OpenAI API 时需要
# openai

# 可选：本地模型 4-bit 量化时需要
# bitsandbytes
